agents modeling/Agents/Supervisor Agent/requirements.txt
langgraph
langchain-core
langchain-community
requests
flask
google-cloud-aiplatform
google-generativeai